% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{add_mailing_contact}
\alias{add_mailing_contact}
\title{Add a contact to a Qualtrics mailing list}
\usage{
add_mailing_contact(
  qualtrics_api_key,
  directoryID,
  mailingListID,
  firstname,
  lastname,
  mail,
  base_url = "https://usf.az1.qualtrics.com"
)
}
\arguments{
\item{qualtrics_api_key}{API token for Qualtrics.}

\item{directoryID}{Directory identifier that contains the mailing list.}

\item{mailingListID}{Mailing list identifier.}

\item{firstname}{Contact first name.}

\item{lastname}{Contact last name.}

\item{mail}{Contact email address.}

\item{base_url}{Character string for Qualtrics base URL.}
}
\value{
Parsed response invisibly.
}
\description{
Add a contact to a Qualtrics mailing list
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
