% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_HelperFunction.R
\name{CheckNotEmptyString}
\alias{CheckNotEmptyString}
\title{Check non-empty string}
\usage{
CheckNotEmptyString(x)
}
\arguments{
\item{x}{Character scalar.}
}
\value{
Logical scalar.
}
\description{
Small utility to validate that a character value is not NULL/NA/empty.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
