% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_HelperClarity.R
\name{clarity_fetch_csv_from_url_through_httr2}
\alias{clarity_fetch_csv_from_url_through_httr2}
\title{Fetch Clarity data from URL through httr2 request for fixture capture}
\usage{
clarity_fetch_csv_from_url_through_httr2(clarityURL, clarityReportId = NULL)
}
\arguments{
\item{clarityURL}{Character string URL to the CSV data file}
}
\value{
Data frame with the downloaded data
}
\description{
Downloads CSV data from a Clarity API-provided URL.
}
\keyword{internal}
