% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_LogicAndFlow.R
\name{create_block_flow_body_qualtrics}
\alias{create_block_flow_body_qualtrics}
\title{Create a block flow element body}
\usage{
create_block_flow_body_qualtrics(blockID, flowID)
}
\arguments{
\item{blockID}{Block ID.}

\item{flowID}{Flow ID.}
}
\value{
A list representing a Qualtrics Block flow element.
}
\description{
Create a block flow element body
}
\concept{role:helper}
