% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_LogicAndFlow.R
\name{create_branch_flow_body_qualtrics}
\alias{create_branch_flow_body_qualtrics}
\title{Create a branch flow element body}
\usage{
create_branch_flow_body_qualtrics(description, flowID, logic, subflow = list())
}
\arguments{
\item{description}{Character description of the branch.}

\item{flowID}{Character/ID for the flow.}

\item{logic}{List representing the branch logic.}

\item{subflow}{Optional subflow list.}
}
\value{
A list representing a Qualtrics Branch flow element.
}
\description{
Create a branch flow element body
}
\concept{role:helper}
