% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_LogicAndFlow.R
\name{create_display_logic_qualtrics}
\alias{create_display_logic_qualtrics}
\title{Create display logic element for Qualtrics}
\usage{
create_display_logic_qualtrics(
  logicType,
  leftOperand,
  operator,
  rightOperand,
  eqType
)
}
\arguments{
\item{logicType}{Character logic type (e.g. "EmbeddedField").}

\item{leftOperand}{Character left operand name.}

\item{operator}{Character comparison operator.}

\item{rightOperand}{Character right operand value.}

\item{eqType}{Character expression type.}
}
\value{
A list representing a Qualtrics display logic BooleanExpression.
}
\description{
Helper to create a single display logic BooleanExpression used in Qualtrics
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
