% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_LogicAndFlow.R
\name{create_email_logic_qualtrics}
\alias{create_email_logic_qualtrics}
\title{Create display logic for recipient email}
\usage{
create_email_logic_qualtrics(emailAddress)
}
\arguments{
\item{emailAddress}{Character email address. If NA, will be coerced to "0".}
}
\value{
A list representing the display logic for email matching.
}
\description{
Shortcut wrapper to create an email-specific display logic where the
left operand is \code{RecipientEmail}.
}
\concept{role:helper}
