% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_LogicAndFlow.R
\name{create_flow_email_logic_qualtrics}
\alias{create_flow_email_logic_qualtrics}
\title{Create survey flow logic for recipient email}
\usage{
create_flow_email_logic_qualtrics(emailAddress)
}
\arguments{
\item{emailAddress}{Character email address. NA becomes "0".}
}
\value{
A list representing the flow logic for email comparison.
}
\description{
Create survey flow logic for recipient email
}
\concept{role:helper}
