% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_mailing_list}
\alias{create_mailing_list}
\title{Create a mailing list}
\usage{
create_mailing_list(qualtricsKey, directoryID, listName)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{directoryID}{Character directory id.}

\item{listName}{Character list name.}
}
\value{
Mailing list id (character).
}
\description{
Create a mailing list
}
\concept{role:helper}
