% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_matrix_body}
\alias{create_matrix_body}
\title{Create body for a matrix question}
\usage{
create_matrix_body(
  questionStatements,
  questionText,
  questionColumns,
  allowMultipleAnswer
)
}
\arguments{
\item{questionStatements}{Character vector of statements (rows).}

\item{questionText}{Character question text.}

\item{questionColumns}{Character vector of column choices.}

\item{allowMultipleAnswer}{Logical whether multiple answers per row allowed.}
}
\value{
A list ready for the Qualtrics API.
}
\description{
Create body for a matrix question
}
\concept{role:helper}
