% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_multiple_choice_body}
\alias{create_multiple_choice_body}
\title{Create body for multiple choice question}
\usage{
create_multiple_choice_body(questionOptions, questionText, allowMultipleChoice)
}
\arguments{
\item{questionOptions}{Character vector of choices.}

\item{questionText}{Character question text.}

\item{allowMultipleChoice}{Logical whether multiple choices are allowed.}
}
\value{
A list ready to be serialized to JSON for the Qualtrics API.
}
\description{
Create body for multiple choice question
}
\concept{role:helper}
