% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_new_qualtrics_question}
\alias{create_new_qualtrics_question}
\title{Create a new question in a survey}
\usage{
create_new_qualtrics_question(qualtricsKey, jsonBody, surveyID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{jsonBody}{List body for the question JSON.}

\item{surveyID}{Character survey id.}
}
\value{
Parsed API response as a list.
}
\description{
Create a new question in a survey
}
\concept{role:helper}
