% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_survey}
\alias{create_survey}
\title{Create a new Qualtrics survey}
\usage{
create_survey(qualtricsKey, surveyName)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyName}{Character name for the new survey.}
}
\value{
Parsed JSON response from the Qualtrics API as a list.
}
\description{
Wrapper for the Qualtrics API to create a new survey definition.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
