% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_text_box_body}
\alias{create_text_box_body}
\title{Build a text box question body}
\usage{
create_text_box_body(questionText, allowMultipleLineTextbox = FALSE)
}
\arguments{
\item{questionText}{Character string with the question prompt.}

\item{allowMultipleLineTextbox}{Logical; if TRUE uses multi-line selector.}
}
\value{
A list suitable for Qualtrics question creation.
}
\description{
Build a text box question body
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
