% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{delete_mailing_list}
\alias{delete_mailing_list}
\title{Delete a mailing list}
\usage{
delete_mailing_list(
  qualtrics_api_key,
  directoryID,
  mailingID,
  base_url = "https://usf.az1.qualtrics.com"
)
}
\arguments{
\item{qualtrics_api_key}{API token for Qualtrics.}

\item{directoryID}{Directory identifier.}

\item{mailingID}{Mailing list identifier to delete.}

\item{base_url}{Character string for Qualtrics base URL.}
}
\value{
Parsed response invisibly.
}
\description{
Delete a mailing list
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
