% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{extract_question_ids}
\alias{extract_question_ids}
\title{Extract question IDs from survey definition}
\usage{
extract_question_ids(qualtricsKey, surveyID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyID}{Character survey id.}
}
\value{
Character vector of question IDs.
}
\description{
Extract question IDs from survey definition
}
\concept{role:helper}
