% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_HelperFunction.R
\name{generate_box_plot_data}
\alias{generate_box_plot_data}
\title{Generate a simple boxplot (deprecated)}
\usage{
generate_box_plot_data(datafiles, columnName, renameName, newLabel, dataInfo)
}
\arguments{
\item{datafiles}{A named list of data frames.}

\item{columnName}{Column name (string) to plot.}

\item{renameName}{Y-axis label.}

\item{newLabel}{Plot title.}

\item{dataInfo}{Monitor information table used for labels.}
}
\value{
A \code{ggplot} object.
}
\description{
Deprecated convenience wrapper that assembles a boxplot from multiple
data sources. Left for backward compatibility.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
