% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{get_all_survey_questions}
\alias{get_all_survey_questions}
\title{Get all questions for a Qualtrics survey}
\usage{
get_all_survey_questions(qualtricsKey, surveyID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyID}{Character survey id.}
}
\value{
Parsed JSON response as a list.
}
\description{
Get all questions for a Qualtrics survey
}
\concept{role:helper}
