% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_main_personnel_list}
\alias{get_main_personnel_list}
\title{Read main personnel CSV and optionally filter by role}
\usage{
get_main_personnel_list(
  role = NULL,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{role}{Optional character or list of roles to filter by.}

\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble of participants, optionally filtered.
}
\description{
Read main personnel CSV and optionally filter by role
}
\concept{role:helper}
