% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_monthly_log}
\alias{get_monthly_log}
\title{Read monthly Qualtrics log CSV}
\usage{
get_monthly_log(root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER"))
}
\arguments{
\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble with the monthly log contents.
}
\description{
Read monthly Qualtrics log CSV
}
\concept{role:helper}
