% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_monthly_question_descriptions}
\alias{get_monthly_question_descriptions}
\title{Read and split monthly question descriptions}
\usage{
get_monthly_question_descriptions(
  questionDescFileName,
  splitLikeUnresponsed = FALSE,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{questionDescFileName}{Character file name under \code{CSV/Qualtrics/Monthly}.}

\item{splitLikeUnresponsed}{Logical toggle for alternative split format.}

\item{root_folder}{Character root folder path.}
}
\value{
A tibble with separated columns.
}
\description{
Read and split monthly question descriptions
}
\concept{role:helper}
