% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_monthly_question_shortlist}
\alias{get_monthly_question_shortlist}
\title{Read monthly question shortlist}
\usage{
get_monthly_question_shortlist(root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER"))
}
\arguments{
\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble with question shortlist.
}
\description{
Read monthly question shortlist
}
\concept{role:helper}
