% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{get_qualtrics_survey}
\alias{get_qualtrics_survey}
\title{Get Qualtrics survey definition}
\usage{
get_qualtrics_survey(qualtricsKey, surveyID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyID}{Character survey id.}
}
\value{
Parsed JSON response as a list.
}
\description{
Get Qualtrics survey definition
}
\concept{role:helper}
