% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_question_descriptions}
\alias{get_question_descriptions}
\title{Read question descriptions (weekly)}
\usage{
get_question_descriptions(
  questionDescFileName,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{questionDescFileName}{Character file name under \code{CSV/Qualtrics/Weekly}.}

\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble with question descriptions and split columns.
}
\description{
Read question descriptions (weekly)
}
\concept{role:helper}
