% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_File.R
\name{get_reference_site_shortcodes}
\alias{get_reference_site_shortcodes}
\title{Get Reference Site Short Codes}
\usage{
get_reference_site_shortcodes()
}
\value{
Character vector of short codes from all reference sites
}
\description{
Extracts short codes from reference sites for identifying co-located sensors.
Used to dynamically determine which sensors are co-located with reference stations.
}
\details{
Reads reference site data and extracts short codes to identify co-located sensors.
This eliminates hardcoded values and ensures accuracy when reference sites change.
}
\examples{
\dontrun{
ref_codes <- get_reference_site_shortcodes()
}

}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
