% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_responsed_analyst_list}
\alias{get_responsed_analyst_list}
\title{Get analysts who have responded (monthly)}
\usage{
get_responsed_analyst_list(
  responseFileName,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{responseFileName}{Character file name of monthly responses.}

\item{root_folder}{Character root folder path.}
}
\value{
A tibble of analysts who have responded.
}
\description{
Get analysts who have responded (monthly)
}
\concept{role:helper}
