% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_responsed_personnel_list}
\alias{get_responsed_personnel_list}
\title{Get personnel who responded (weekly)}
\usage{
get_responsed_personnel_list(
  responseFileName,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{responseFileName}{Character. Filename in \verb{CSV/Qualtrics/Weekly/} with responses.}

\item{root_folder}{Character. Root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble of personnel (with sensor rows) who responded.
}
\description{
Get personnel who responded (weekly)
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
