% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{get_survey_flow}
\alias{get_survey_flow}
\title{Get survey flow}
\usage{
get_survey_flow(qualtricsKey, surveyID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyID}{Character survey id.}
}
\value{
Parsed API response as a list.
}
\description{
Get survey flow
}
\concept{role:helper}
