% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_unresolved_monitor_log}
\alias{get_unresolved_monitor_log}
\title{Read unresolved monitor CSV}
\usage{
get_unresolved_monitor_log(root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER"))
}
\arguments{
\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble with unresolved monitors.
}
\description{
Read unresolved monitor CSV
}
\concept{role:helper}
