% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_unresponsed_personnel_list}
\alias{get_unresponsed_personnel_list}
\title{Get personnel who have not responded (weekly)}
\usage{
get_unresponsed_personnel_list(
  responseFileName,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{responseFileName}{Character file name of weekly responses.}

\item{root_folder}{Character root folder path.}
}
\value{
A tibble of personnel who have not responded.
}
\description{
Get personnel who have not responded (weekly)
}
\concept{role:helper}
