% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_weekly_log}
\alias{get_weekly_log}
\title{Read weekly Qualtrics log CSV}
\usage{
get_weekly_log(root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER"))
}
\arguments{
\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble with the weekly log contents.
}
\description{
Read weekly Qualtrics log CSV
}
\concept{role:helper}
