% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_weekly_personnel_list}
\alias{get_weekly_personnel_list}
\title{Read weekly personnel list from monitor tracking}
\usage{
get_weekly_personnel_list(root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER"))
}
\arguments{
\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble of personnel and short codes.
}
\description{
Reads the QATimeshift Excel sheet and extracts Name/Email/SiteShortCode.
}
\concept{role:helper}
