% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_weekly_question_shortlist}
\alias{get_weekly_question_shortlist}
\title{Read weekly question shortlist}
\usage{
get_weekly_question_shortlist(root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER"))
}
\arguments{
\item{root_folder}{Character. Root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble with the weekly question shortlist.
}
\description{
Read weekly question shortlist
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
