% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_HelperFunction.R
\name{GetAnnotateInfoForScatterplot}
\alias{GetAnnotateInfoForScatterplot}
\title{Annotation values for scatterplots}
\usage{
GetAnnotateInfoForScatterplot(type)
}
\arguments{
\item{type}{Character: one of "PM25" or "NO2".}
}
\value{
A named list with \code{hourlyLabel}, \code{annualLabel}, \code{hourlyValue}, and \code{annualValue}.
}
\description{
Return plotting annotation values (labels and numeric thresholds) for
common pollutants used in plots.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
