% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{publish_survey}
\alias{publish_survey}
\title{Publish a survey version}
\usage{
publish_survey(qualtricsKey, surveyID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyID}{Character survey id.}
}
\value{
Parsed API response as a list.
}
\description{
Publish a survey version
}
\concept{role:helper}
