% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{set_qualtrics_survey_active}
\alias{set_qualtrics_survey_active}
\title{Set survey active flag}
\usage{
set_qualtrics_survey_active(qualtricsKey, surveyID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyID}{Character survey id.}
}
\value{
Parsed API response as a list (invisible).
}
\description{
Set survey active flag
}
\concept{role:helper}
