% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{write_to_monthly_template_update_log}
\alias{write_to_monthly_template_update_log}
\title{Alias for weekly template update log writer}
\usage{
write_to_monthly_template_update_log(
  originDate,
  neededAction,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{originDate}{Value to write to \code{OriginDate} column.}

\item{neededAction}{Value to write to \code{Action} column.}

\item{root_folder}{Character. Root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
NULL, called for side effect.
}
\description{
Alias for weekly template update log writer
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
