% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{add_mailing_contact}
\alias{add_mailing_contact}
\title{Add a contact to a mailing list}
\usage{
add_mailing_contact(
  qualtricsKey,
  directoryID,
  mailingListID,
  firstname,
  lastname,
  mail
)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{directoryID}{Character directory id.}

\item{mailingListID}{Character mailing list id.}

\item{firstname}{Character first name.}

\item{lastname}{Character last name.}

\item{mail}{Character email address.}
}
\value{
Parsed API response as a list (invisible).
}
\description{
Add a contact to a mailing list
}
\concept{role:helper}
