% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{check_exist_in_log}
\alias{check_exist_in_log}
\title{Check existence in log}
\usage{
check_exist_in_log(logFile, originDate, neededAction)
}
\value{
Logical TRUE if a matching row exists, otherwise FALSE.
}
\description{
Check existence in log
}
\concept{role:helper}
