% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{check_not_exist_in_log}
\alias{check_not_exist_in_log}
\title{Check not exist in log}
\usage{
check_not_exist_in_log(logFile, originDate, neededAction)
}
\arguments{
\item{logFile}{Tibble or data.frame log file.}

\item{originDate}{Character or date to match in \code{OriginDate}.}

\item{neededAction}{Character to match in \code{Action}.}
}
\value{
Logical TRUE if no matching row exists, otherwise FALSE.
}
\description{
Check not exist in log
}
\concept{role:helper}
