% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Monthly_UpdateTemplateSurvey.R
\name{copy_original_QAMonitorFile}
\alias{copy_original_QAMonitorFile}
\title{Copy the original MonitorTracking script for QA purposes}
\usage{
copy_original_QAMonitorFile()
}
\value{
None. The function performs file operations and does not return a value.
}
\description{
This function copies the original CAMNMonitorTracking.xlsx file to a specified subfolder with a new name that includes the current month and year.
}
\details{
The function checks if the target subfolder exists and creates it if necessary. It then constructs a new file name based on the current date and copies the original file to the new location with the new name.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_monthly_helper}
