% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_LogicAndFlow.R
\name{create_flow_logic_qualtrics}
\alias{create_flow_logic_qualtrics}
\title{Create a survey flow logic element for Qualtrics}
\usage{
create_flow_logic_qualtrics(
  logicType,
  leftOperand,
  operator,
  rightOperand,
  eqType
)
}
\arguments{
\item{logicType}{Character logic type.}

\item{leftOperand}{Character left operand.}

\item{operator}{Character operator.}

\item{rightOperand}{Character right operand.}

\item{eqType}{Character expression type.}
}
\value{
A list representing a survey flow BooleanExpression.
}
\description{
Creates a BooleanExpression list suitable for Qualtrics survey flow rules.
}
\concept{role:helper}
