% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_survey}
\alias{create_survey}
\title{Create a new Qualtrics survey}
\usage{
create_survey(
  qualtrics_api_key,
  surveyName,
  base_url = "https://usf.az1.qualtrics.com"
)
}
\arguments{
\item{qualtrics_api_key}{API token for Qualtrics.}

\item{surveyName}{Character string with the desired survey name.}

\item{base_url}{Character string for Qualtrics base URL. Defaults to the USF datacenter.}
}
\value{
Parsed JSON response from Qualtrics as a list.
}
\description{
Create a new survey resource in Qualtrics using the survey-definitions endpoint.
}
\examples{
\dontrun{
create_survey("<REDACTED_API_KEY>", "My Survey")
}

}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
