% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_survey_qsf}
\alias{create_survey_qsf}
\title{Create a new survey from QSF data}
\usage{
create_survey_qsf(qualtricsKey, qsfData)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{qsfData}{List representing the QSF JSON.}
}
\value{
New survey ID as character.
}
\description{
Create a new survey from QSF data
}
\concept{role:helper}
