% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_text_box_body}
\alias{create_text_box_body}
\title{Create body for a text box question}
\usage{
create_text_box_body(questionText, allowMultipleLineTextbox)
}
\arguments{
\item{questionText}{Character question text.}

\item{allowMultipleLineTextbox}{Logical whether multi-line input is allowed.}
}
\value{
A list ready for the Qualtrics API.
}
\description{
Create body for a text box question
}
\concept{role:helper}
