% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{create_text_form_body}
\alias{create_text_form_body}
\title{Create body for a text form question}
\usage{
create_text_form_body(questionOptions, questionText)
}
\arguments{
\item{questionOptions}{Character vector for form fields.}

\item{questionText}{Character question text.}
}
\value{
A list ready for the Qualtrics API.
}
\description{
Create body for a text form question
}
\concept{role:helper}
