% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{delete_mailing_list}
\alias{delete_mailing_list}
\title{Delete a mailing list}
\usage{
delete_mailing_list(qualtricsKey, directoryID, mailingID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{directoryID}{Character directory id.}

\item{mailingID}{Character mailing list id.}
}
\value{
Parsed API response as a list (invisible).
}
\description{
Delete a mailing list
}
\concept{role:helper}
