% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_first_save_data_from_monthly_log}
\alias{get_first_save_data_from_monthly_log}
\title{Alias for monthly log lookup}
\usage{
get_first_save_data_from_monthly_log(logFile, originDate, neededAction)
}
\arguments{
\item{logFile}{A data.frame or tibble representing the log.}

\item{originDate}{Value to match in \code{OriginDate} column.}

\item{neededAction}{Value to match in \code{Action} column.}
}
\value{
The matched \code{SaveData} value or \code{NULL}.
}
\description{
Wrapper around \code{get_first_save_data_from_weekly_log} for monthly logs.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
