% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_first_save_data_from_weekly_log}
\alias{get_first_save_data_from_weekly_log}
\title{Get first SaveData value from weekly log}
\usage{
get_first_save_data_from_weekly_log(logFile, originDate, neededAction)
}
\arguments{
\item{logFile}{A data.frame or tibble representing the log.}

\item{originDate}{Value to match in \code{OriginDate} column.}

\item{neededAction}{Value to match in \code{Action} column.}
}
\value{
The first \code{SaveData} value matching the filters, or \code{NULL} if none.
}
\description{
Get first SaveData value from weekly log
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
