% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_monitor_sites}
\alias{get_monitor_sites}
\title{Read monitor sites from timeshift Excel}
\usage{
get_monitor_sites(root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER"))
}
\arguments{
\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble of monitor site metadata.
}
\description{
Read monitor sites from timeshift Excel
}
\concept{role:helper}
