% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_monthly_question_info}
\alias{get_monthly_question_info}
\title{Read monthly question info CSV}
\usage{
get_monthly_question_info(root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER"))
}
\arguments{
\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
Data.frame of the question info.
}
\description{
Read monthly question info CSV
}
\concept{role:helper}
