% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_monthly_responses_list}
\alias{get_monthly_responses_list}
\title{Read processed monthly responses}
\usage{
get_monthly_responses_list(
  responseFileName,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{responseFileName}{Character file name under \code{CSV/Qualtrics/Monthly}.}

\item{root_folder}{Character root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble of responses.
}
\description{
Read processed monthly responses
}
\concept{role:helper}
