% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{get_question_descriptions}
\alias{get_question_descriptions}
\title{Read and parse question descriptions for weekly surveys}
\usage{
get_question_descriptions(
  questionDescFileName,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\arguments{
\item{questionDescFileName}{Character. Filename in \verb{CSV/Qualtrics/Weekly/}.}

\item{root_folder}{Character. Root folder path. Defaults to \code{Sys.getenv("UPLOAD_ROOT_FOLDER")}.}
}
\value{
A tibble with separated \code{DeviceID}, \code{DeviceType}, and \code{Site} columns.
}
\description{
Read and parse question descriptions for weekly surveys
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
