% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{get_question_info_qualtrics}
\alias{get_question_info_qualtrics}
\title{Extract question text and export tags}
\usage{
get_question_info_qualtrics(qualtricsKey, surveyID)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyID}{Character survey id.}
}
\value{
A tibble with \code{QuestionText} and \code{DataExportTag}.
}
\description{
Returns a tibble with \code{QuestionText} and \code{DataExportTag} for each question
in the provided survey.
}
\concept{role:helper}
