% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{get_survey_flow}
\alias{get_survey_flow}
\title{Get survey flow}
\usage{
get_survey_flow(
  qualtrics_api_key,
  surveyID,
  base_url = "https://usf.az1.qualtrics.com"
)
}
\arguments{
\item{qualtrics_api_key}{API token for Qualtrics.}

\item{surveyID}{Survey identifier.}

\item{base_url}{Character string for Qualtrics base URL.}
}
\value{
Parsed JSON response as a list.
}
\description{
Retrieve the flow object (survey structure) for a given survey.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
