% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{get_survey_qsf}
\alias{get_survey_qsf}
\title{Get QSF (full survey JSON) for a survey}
\usage{
get_survey_qsf(qualtricsKey, surveyID, newSurveyName)
}
\arguments{
\item{qualtricsKey}{Character API token.}

\item{surveyID}{Character survey id.}

\item{newSurveyName}{Character new survey name to assign to the copy data.}
}
\value{
The \code{result} field of the QSF response as a list.
}
\description{
Get QSF (full survey JSON) for a survey
}
\concept{role:helper}
