% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{publish_survey}
\alias{publish_survey}
\title{Publish a Qualtrics survey version}
\usage{
publish_survey(
  qualtrics_api_key,
  surveyID,
  base_url = "https://usf.az1.qualtrics.com"
)
}
\arguments{
\item{qualtrics_api_key}{API token for Qualtrics.}

\item{surveyID}{Survey identifier.}

\item{base_url}{Character string for Qualtrics base URL.}
}
\value{
Parsed JSON response as a list.
}
\description{
Marks a new survey version as published.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
