% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Monthly_GenerateReport.R
\name{qualtrics_concentrate_monthly_data_from_responses}
\alias{qualtrics_concentrate_monthly_data_from_responses}
\title{Qualtrics monthly - concentrate data from survey responses}
\usage{
qualtrics_concentrate_monthly_data_from_responses()
}
\value{
None
}
\description{
Qualtrics monthly - concentrate data from survey responses
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:false}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_monthly}
