% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Monthly_CreateSurvey.R
\name{qualtrics_create_monthly_survey}
\alias{qualtrics_create_monthly_survey}
\title{Create monthly QA surveys through Qualtrics API}
\usage{
qualtrics_create_monthly_survey()
}
\value{
None
}
\description{
Create monthly QA surveys through Qualtrics API
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:false}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_monthly}
