% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Weekly_CreateSurvey.R
\name{qualtrics_create_weekly_survey}
\alias{qualtrics_create_weekly_survey}
\title{Create Weekly Survey in Qualtrics API}
\usage{
qualtrics_create_weekly_survey()
}
\value{
None
}
\description{
This function creates a new weekly survey in Qualtrics using a predefined template. It checks if the survey for the current week has already been created to avoid duplicates.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:false}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_weekly}
