% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Weekly_GenerateReport.R
\name{qualtrics_generate_weekly_report}
\alias{qualtrics_generate_weekly_report}
\title{Qualtrics Weekly - concentrate data from responses}
\usage{
qualtrics_generate_weekly_report()
}
\description{
This function concentrates data from survey responses for weekly report generation.
It checks if the report has already been sent for the current week, retrieves response data,
merges it with personnel information, and prepares the data for report generation.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_weekly}
