% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Monthly_SendSurvey.R
\name{qualtrics_send_monthly_survey}
\alias{qualtrics_send_monthly_survey}
\title{Send Monthly Survey via Qualtrics API}
\usage{
qualtrics_send_monthly_survey()
}
\value{
None
}
\description{
Send Monthly Survey via Qualtrics API
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:false}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_monthly}
