% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Weekly_SendReminder.R
\name{qualtrics_send_reminder_email}
\alias{qualtrics_send_reminder_email}
\title{Qualtrics Weekly - send survey reminder through emails}
\usage{
qualtrics_send_reminder_email()
}
\description{
Qualtrics Weekly - send survey reminder through emails
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:false}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_weekly}
