% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Weekly_SendSurvey.R
\name{qualtrics_send_weekly_survey}
\alias{qualtrics_send_weekly_survey}
\title{Qualtrics Weekly - send weekly survey through emails}
\usage{
qualtrics_send_weekly_survey()
}
\description{
Qualtrics Weekly - send weekly survey through emails
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:false}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_weekly}
