% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_PurpleAir.R
\name{save_purpleAir_to_csv}
\alias{save_purpleAir_to_csv}
\title{Download from PurpleAir API, process, and save to CSV files}
\usage{
save_purpleAir_to_csv(current_date)
}
\arguments{
\item{current_date}{Date to determine the previous month for data extraction}
}
\value{
None. Saves processed data to CSV files in specified directories.
}
\description{
Download from PurpleAir API, process, and save to CSV files
}
\examples{
Example usage:
myDate <- Sys.Date()
save_purpleAir_to_csv(myDate)
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:false}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:download}
