% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Monthly_UpdateTemplateSurvey.R
\name{update_list_unresolved_monitor_qualtrics}
\alias{update_list_unresolved_monitor_qualtrics}
\title{Get the monthly question shortlist from a CSV file}
\usage{
update_list_unresolved_monitor_qualtrics(
  qualtKey,
  surveyId,
  unresolveQuestionID
)
}
\value{
A tibble containing the question IDs and sensor types from the CSV file.
}
\description{
This function reads a CSV file containing a shortlist of questions for monthly updates and returns it as a tibble.
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:false}
\concept{cleanupParameters:false}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:qualtrics_monthly_helper}
