% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Main.R
\name{update_survey_flow_element}
\alias{update_survey_flow_element}
\title{Update a survey flow element}
\usage{
update_survey_flow_element(
  qualtrics_api_key,
  surveyID,
  flowID,
  jsonBody,
  base_url = "https://usf.az1.qualtrics.com"
)
}
\arguments{
\item{qualtrics_api_key}{API token for Qualtrics.}

\item{surveyID}{Survey identifier.}

\item{flowID}{Flow element identifier.}

\item{jsonBody}{A list payload for the flow element update.}

\item{base_url}{Character string for Qualtrics base URL.}
}
\value{
Parsed response invisibly.
}
\description{
Update a survey flow element
}
\concept{addRoxygenComments:true}
\concept{cleanupComments:true}
\concept{cleanupDependenciesNamespace:true}
\concept{cleanupParameters:true}
\concept{removedDependencies:true}
\concept{removedRawFunctionCalls:true}
\concept{removedSensitiveInfo:true}
\concept{role:helper}
