% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userSetupHelper.R
\name{validate_input}
\alias{validate_input}
\title{User Setup Helper for RCommunityAirSensor Package}
\usage{
validate_input(prompt, allow_empty = FALSE, variable_name = "")
}
\description{
This file contains functions to help users set up their environment variables
and folder structures for the RCommunityAirSensor package.
}
