% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualtrics_Helper_Secondary.R
\name{write_to_monthly_template_update_log}
\alias{write_to_monthly_template_update_log}
\title{Append a row to the monthly template update log (alias)}
\usage{
write_to_monthly_template_update_log(
  originDate,
  neededAction,
  root_folder = Sys.getenv("UPLOAD_ROOT_FOLDER")
)
}
\description{
Append a row to the monthly template update log (alias)
}
\concept{role:helper}
